+++ contracts/script/multisigTransactionProposals/safeSDK/upgradeProxy.ts
@@ -3,7 +3,7 @@ import { ethers } from "ethers";
 import { EthersAdapter } from "@safe-global/protocol-kit";
 import SafeApiKit from "@safe-global/api-kit";
 import Safe from "@safe-global/protocol-kit";
-import { getEnvVar, createSafeTransactionData, validateEthereumAddress } from "./utils";
+import { getEnvVar, createSafeTransaction, validateEthereumAddress, createAndSignSafeTransaction } from "./utils";
 const UPGRADE_PROXY_CMD = "upgradeProxy" as const;
 
 // declaring the type returned by the createTransaction method in the safe package locally (since the return type isn't exposed) so that if it's updated, it's reflected here too
@@ -91,16 +91,19 @@ export async function proposeUpgradeTransaction(
   upgradeData: UpgradeData,
 ) {
   // Prepare the transaction data to upgrade the proxy
-  let data = createUpgradeTxData(upgradeData.implementationAddress, upgradeData.initData);
+  const abi = ["function upgradeToAndCall(address,bytes)"];
+  // Encode the function call with the new implementation address and its init data
+  const data = new ethers.Interface(abi).encodeFunctionData("upgradeToAndCall", [
+    upgradeData.implementationAddress,
+    upgradeData.initData,
+  ]);
 
-  // Create the Safe Transaction Object
-  const safeTransaction = await createSafeTransaction(safeSDK, upgradeData.proxyAddress, data, "0");
-
-  // Get the transaction hash and sign the transaction
-  const safeTxHash = await safeSDK.getTransactionHash(safeTransaction);
-
-  // Sign the transaction with orchestrator signer that was specified when we created the safeSDK
-  const senderSignature = await safeSDK.signHash(safeTxHash);
+  // Create & Sign the Safe Transaction Object
+  const { safeTransaction, safeTxHash, senderSignature } = await createAndSignSafeTransaction(
+    safeSDK,
+    upgradeData.proxyAddress,
+    data,
+  );
 
   // Propose the transaction which can be signed by other owners via the Safe UI
   await safeService.proposeTransaction({
@@ -110,45 +113,6 @@ export async function proposeUpgradeTransaction(
     senderAddress: signerAddress,
     senderSignature: senderSignature.data,
   });
-}
-
-/**
- * Function to create the transaction data for setting the new implementation
- * @param {string} newContractAddress - The address of the new implementation
- * @param {string} initData - The initialization data for the new implementation
- * @returns {string} - Encoded transaction data
- */
-function createUpgradeTxData(newContractAddress: string, initData: string): string {
-  // Define the ABI of the function to be called
-  const abi = ["function upgradeToAndCall(address,bytes)"];
-
-  // Encode the function call with the new implementation address and its init data
-  const data = new ethers.Interface(abi).encodeFunctionData("upgradeToAndCall", [newContractAddress, initData]);
-  return data; // Return the encoded transaction data
-}
-
-/**
- * Creates a Safe transaction object
- *
- * @param {Safe} safeSDK - An instance of the Safe SDK
- * @param {string} contractAddress - The address of the contract to interact with
- * @param {string} data - The data payload for the transaction
- * @param {string} value - The value to be sent with the transaction
- * @returns {Promise<any>} - A promise that resolves to the Safe transaction object
- */
-async function createSafeTransaction(
-  safeSDK: Safe,
-  contractAddress: string,
-  data: string,
-  value: string,
-): Promise<LocalSafeTransaction> {
-  // Prepare the safe transaction data with the contract address, data, and value
-  let safeTransactionData = createSafeTransactionData(contractAddress, data, value);
-
-  // Create the safe transaction using the Safe SDK
-  const safeTransaction = await safeSDK.createTransaction({ transactions: [safeTransactionData] });
-
-  return safeTransaction;
 }
 
 main();
